+++ service-base/src/main/java/org/eclipse/hono/service/monitoring/ConnectionEventProducer.java
@@ -16,6 +16,7 @@ package org.eclipse.hono.service.monitoring;
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
 import org.eclipse.hono.auth.Device;
+import org.eclipse.hono.client.HonoClient;
 
 /**
  * Produces connection events.
@@ -23,21 +24,47 @@ import org.eclipse.hono.auth.Device;
  * The interface is intended to be implemented by setups which are interested in receiving technical connection events.
  * Which might not make sense for all protocols adapters. But mostly for connection oriented ones like MQTT or AMQP 1.0.
  * <p>
- * The protocol adapters may call {@link #connected(String, String, Device, JsonObject)} and
- * {@link #disconnected(String, String, Device, JsonObject)} as they see fit. The whole process is a "best effort"
- * process and intended to be used for monitoring/debugging.
+ * The protocol adapters may call {@link #connected(Context, String, String, Device, JsonObject)} and
+ * {@link #disconnected(Context, String, String, Device, JsonObject)} as they see fit. The whole process is a "best
+ * effort" process and intended to be used for monitoring/debugging.
  * <p>
  * When a protocol adapter calls into this producer it must provide some kind of connection ID, which might have a
  * different meaning for different protocol adapters. This should be as identifiable as possible but there is no
  * requirement for this to be unique. However for each connection the protocol adapter must call
- * {@link #disconnected(String, String, Device, JsonObject)} with the same information as it called
- * {@link #connected(String, String, Device, JsonObject)}.
+ * {@link #disconnected(Context, String, String, Device, JsonObject)} with the same information as it called
+ * {@link #connected(Context, String, String, Device, JsonObject)}.
  */
 public interface ConnectionEventProducer {
 
     /**
+     * Context object for the connection events producers.
+     * <p>
+     * This context is only valid for one call and its values must not be stored by the implementation of the
+     * <em>connection event producer</em>.
+     */
+    interface Context {
+
+        /**
+         * Provide the device registry service client provided by the caller.
+         * 
+         * @return The instance of the registry client which the {@link ConnectionEventProducer} method should use. This
+         *         client has to be initialized and started.
+         */
+        HonoClient getDeviceRegistryClient();
+
+        /**
+         * Provide the device message sender client provided by the caller.
+         * 
+         * @return The instance of the message sender client which the {@link ConnectionEventProducer} method should
+         *         use. This client has to be initialized and started.
+         */
+        HonoClient getMessageSenderClient();
+    }
+
+    /**
      * Produce an event for a new connection.
      * 
+     * @param context Protocol adapter context.
      * @param remoteId The ID of the remote endpoint which connected (e.g. a remote address, port, client id, ...).
      * @param protocolAdapter The name of the protocol adapter sending this event. Must not be {@code null}.
      * @param authenticatedDevice The optional authenticated device associated with the connection. May be {@code null}
@@ -46,13 +73,15 @@ public interface ConnectionEventProducer {
      * @return A future which indicates the result of the event production
      * @throws NullPointerException If either the remote ID or the protocol adapter argument are {@code null}.
      */
-    Future<?> connected(String remoteId, String protocolAdapter, Device authenticatedDevice, JsonObject data);
+    Future<?> connected(Context context, String remoteId, String protocolAdapter, Device authenticatedDevice,
+            JsonObject data);
 
     /**
      * Produce an event for a closed connection.
      * 
+     * @param context Protocol adapter context.
      * @param remoteId The ID of the remote endpoint which disconnected. The same ID used in the call to
-     *            {@link #connected(String, String, Device, JsonObject)}
+     *            {@link #connected(Context, String, String, Device, JsonObject)}
      * @param protocolAdapter The name of the protocol adapter sending this event. Must not be {@code null}.
      * @param authenticatedDevice The optional authenticated device associated with the connection. May be {@code null}
      *            if the connection is from an unauthenticated device.
@@ -60,5 +89,6 @@ public interface ConnectionEventProducer {
      * @return A future which indicates the result of the event production
      * @throws NullPointerException If either the remote ID or the protocol adapter argument are {@code null}.
      */
-    Future<?> disconnected(String remoteId, String protocolAdapter, Device authenticatedDevice, JsonObject data);
+    Future<?> disconnected(Context context, String remoteId, String protocolAdapter, Device authenticatedDevice,
+            JsonObject data);
 }